% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak_z.R
\name{liptak.Z}
\alias{liptak.Z}
\title{Liptak's Weighted Z}
\usage{
liptak.Z(
  x,
  y = NULL,
  null.mu = 0,
  alternative = "two.sided",
  var.equal = NULL,
  var.alpha = 0.2
)
}
\arguments{
\item{x}{A numeric vector, or 2xn (or nx2) matrix. If \code{x} is a vector,
then data should be paired with \code{y} by index. If \code{x} is a matrix,
then pairing is done within column. The data must contain at least two
pairwise complete observations, and at least four pairwise incomplete
\code{(NA, !NA)} observations with at least two of those incomplete pairs
having NA in \code{x}.}

\item{y}{A numeric vector. Must contain at least two NAs paired with
non-missing values in \code{x}. For more information on pairwise
requirements, see description of \code{x}. Ignored if \code{x} is a matrix.}

\item{null.mu}{Number indicating the true difference of means.}

\item{alternative}{The alternative hypothesis, one of \code{"two.sided"},
\code{"greater"}, or \code{"less"}}

\item{var.equal}{Logical indicating whether the variance is equal for
testing difference of means. Default \code{NULL} will run a test for
equality of variances.
@param alpha Level at which to reject the assumption of equal variance.
Ignored if \code{var.equal} is not \code{NULL}.}
}
\value{
A list with class \code{"htest"} used for internal purposes to produce
printout.
}
\description{
Method for testing difference of means in partially paired data based on
p-value pooling meta-analysis as proposed by Liptak [1958], with p-values
obtained by t-test.
}
\examples{
# growth in pine trees between 3 and 5 years of age
x <- Loblolly[Loblolly$age == 3, "height"]
y <- Loblolly[Loblolly$age == 5, "height"]
# assign NA to first five obs of age 3, x, and last three obs of age 5, y
x[1:5] <- NA
y[11:14] <- NA

# we expect a difference but test to make sure it exists
liptak.Z(x, y)

# we can also pass x and y as a matrix
X <- cbind(x, y)
liptak.Z(X)

}
