% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lj_zcorr.R
\name{LJ.Zcorr}
\alias{LJ.Zcorr}
\title{Looney and Jones’s Corrected Z-Test}
\usage{
LJ.Zcorr(x, y = NULL, alternative = "two.sided")
}
\arguments{
\item{x}{A numeric vector or 2xn matrix. If \code{x} is a vector, then
data should be paired with \code{y} by index. If \code{x} is a matrix,
then pairing is done within column.}

\item{y}{A numeric vector. Ignored if \code{x} is a matrix.}

\item{alternative}{The alternative hypothesis, one of \code{"two.sided"},
\code{"greater"}, or \code{"less"}}
}
\value{
A list with class \code{"htest"} used for internal purposes to produce
printout.
}
\description{
Method for testing difference of means in partially paired data using
correlation between paired observations to correct variance as proposed by
Looney and Jones [2003].
}
\examples{
# growth in orange trees between 118 and 664 days of age
x <- Orange[Orange$age == 118, "circumference"]
y <- Orange[Orange$age == 664, "circumference"]
# assign NA to first tree at age 118, x, and last tree at age 664, y
x[1] <- NA
y[5] <- NA

# we expect a difference but test to make sure it exists
LJ.Zcorr(x, y)

# we can also pass x and y as a matrix
X <- cbind(x, y)
LJ.Zcorr(X)

# Finally, we don't require every group be represented. For example,
# if there are no unpaired x obs:
x <- Orange[Orange$age == 118, "circumference"]
# then the test is still valid:
LJ.Zcorr(x, y)

}
